third_party/WebKit/LayoutTests/http/tests/security/contentSecurityPolicy/1.1/scripthash-allowed.html

<!DOCTYPE html>
<html>
    <head>
        <meta http-equiv="Content-Security-Policy" content="script-src 'sha1-Au4uYFbkf7OYd+ACMnKq96FN3qo='">
        <script>
            if (window.testRunner)
                testRunner.dumpAsText();
            alert('PASS (1/1)');
        </script>
    </head>
    <body>
        <p>
            This tests the effect of a valid script-hash value. It passes if no
            console warning is visible, and the alert() is executed.
        </p>
    </body>
</html>
