third_party/WebKit/LayoutTests/http/tests/security/contentSecurityPolicy/1.1/scripthash-allowed-expected.txt

ALERT: PASS (1/1)
This tests the effect of a valid script-hash value. It passes if no console warning is visible, and the alert() is executed.
